Some testing. Added CHANGES.txt
